NewtonSoftJsonProvider as intermediary to Newtonsoft.Json library

+++ ComputersFactory/ComputersFactory.Data.FileSystem/FileSystemService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 
 using ComputersFactory.Data.FileSystem.Contracts;
@@ -9,12 +10,41 @@ namespace ComputersFactory.Data.FileSystem
     {
         public IEnumerable<string> ReadFromFile(string filePath)
         {
+            if (string.IsNullOrEmpty(filePath))
+            {
+                throw new ArgumentNullException(nameof(filePath));
+            }
+
+            var fileExists = File.Exists(filePath);
+            if (!fileExists)
+            {
+                throw new FileNotFoundException(filePath);
+            }
+
             var content = File.ReadAllLines(filePath);
             return content;
         }
 
         public void WriteToFile(string filePath, string content)
         {
+            if (string.IsNullOrEmpty(content))
+            {
+                throw new ArgumentNullException(nameof(content));
+            }
+
+            if (string.IsNullOrEmpty(filePath))
+            {
+                throw new ArgumentNullException(nameof(filePath));
+            }
+
+            var filePathInfo = new FileInfo(filePath);
+            var directoryName = filePathInfo.DirectoryName;
+            var containingDirectoryExists = Directory.Exists(directoryName);
+            if (!containingDirectoryExists)
+            {
+                throw new DirectoryNotFoundException(filePath);
+            }
+
             File.WriteAllText(filePath, content);
         }
     }
